# openapi.yaml
...
openapi: 3.1.0
info:
  title: photolib
  description: "\nA local, private photo library. Natural-language search over your photos,\nface search\
    \ and recognition, and date/location browsing \u2014 all running on your\nown machine with no network\
    \ calls.\n"
  version: 2.0.0
paths:
...
  /api/v1/images/trash:
    post:
      tags:
      - images
      summary: Trash Images
      description: "Move photos to the OS Recycle Bin and drop them from the library.\n\nNothing is permanently\
        \ deleted \u2014 the files are recoverable from the\nRecycle Bin. Rows whose files are already\
        \ missing are cleaned up too."
      operationId: trash_images_api_v1_images_trash_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrashRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/export:
    post:
      tags:
      - images
      summary: Export Images
      description: 'Copy originals into a folder of the user''s choosing.


        Purely additive: sources are never moved, renamed, or overwritten.'
      operationId: export_images_api_v1_images_export_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/{image_id}/reveal:
    post:
      tags:
      - images
      summary: Reveal Image
      description: Open the OS file manager with this file selected. Local desktop only.
      operationId: reveal_image_api_v1_images__image_id__reveal_post
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/{image_id}:
    get:
      tags:
      - images
      summary: Get Original
      description: The original file, straight off disk.
      operationId: get_original_api_v1_images__image_id__get
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      - name: download
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Download
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/{image_id}/thumbnail:
    get:
      tags:
      - images
      summary: Get Thumbnail
      description: A cached thumbnail, generated on first request.
      operationId: get_thumbnail_api_v1_images__image_id__thumbnail_get
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      - name: size
        in: query
        required: false
        schema:
          type: string
          enum:
          - small
          - medium
          - grid
          - large
          - preview
          default: grid
          title: Size
      - name: format
        in: query
        required: false
        schema:
          type: string
          enum:
          - webp
          - jpeg
          default: webp
          title: Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/{image_id}/details:
    get:
      tags:
      - images
      summary: Get Details
      description: Full metadata, the people in the photo, and every detected face box.
      operationId: get_details_api_v1_images__image_id__details_get
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/{image_id}/text:
    get:
      tags:
      - images
      summary: Get Text
      description: The full text found inside the image (details carries an excerpt).
      operationId: get_text_api_v1_images__image_id__text_get
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/{image_id}/similar:
    get:
      tags:
      - images
      summary: Get Similar
      description: Visually similar photos, using the stored image embedding.
      operationId: get_similar_api_v1_images__image_id__similar_get
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 24
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/{image_id}/faces:
    get:
      tags:
      - images
      summary: Get Faces
      operationId: get_faces_api_v1_images__image_id__faces_get
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ExportRequest:
      properties:
        image_ids:
          items:
            type: integer
          type: array
          maxItems: 500
          minItems: 1
          title: Image Ids
        folder:
          type: string
          minLength: 1
          title: Folder
      type: object
      required:
      - image_ids
      - folder
      title: ExportRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TrashRequest:
      properties:
        image_ids:
          items:
            type: integer
          type: array
          maxItems: 500
          minItems: 1
          title: Image Ids
      type: object
      required:
      - image_ids
      title: TrashRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError